don't source /etc/sysconfig/etherpad here
We do that in the update script that *calls* this, and that update script also works around a bug caused by mismatched paths. Seems nicer to leave the fix there.

That path could possibly use $ETHERPADDIR as defined here, but I'm not entirely sure that will work on the end system and don't have the time to test properly.